requirements-dev.txt
# Development dependencies for AWS Research Wizard
# Install with: pip install -r requirements-dev.txt

# Core dependencies
boto3>=1.28.0
botocore>=1.31.0
streamlit>=1.25.0
pandas>=2.0.0
numpy>=1.25.0
plotly>=5.15.0
pyyaml>=6.0
requests>=2.31.0
click>=8.1.0
rich>=13.0.0
jinja2>=3.1.0
typing-extensions>=4.5.0

# Code quality and linting
black>=23.7.0
isort>=5.12.0
flake8>=6.0.0
flake8-docstrings>=1.7.0
flake8-bugbear>=23.7.10
flake8-comprehensions>=3.14.0
flake8-simplify>=0.20.0
flake8-pytest-style>=1.7.2
mypy>=1.5.0
pylint>=2.17.0
bandit>=1.7.5
autoflake>=2.2.0
pyupgrade>=3.10.1
docformatter>=1.7.5

# Testing framework
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-asyncio>=0.21.0
pytest-xdist>=3.3.0
pytest-timeout>=2.1.0
pytest-randomly>=3.13.0
coverage>=7.3.0
moto>=4.2.0
responses>=0.23.3

# Type checking stubs
types-PyYAML>=6.0.0
types-requests>=2.31.0
boto3-stubs[essential]>=1.28.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0
sphinx-autodoc-typehints>=1.24.0
myst-parser>=2.0.0
sphinx-click>=5.0.0

# Pre-commit hooks
pre-commit>=3.4.0

# Development utilities
tox>=4.11.0
wheel>=0.41.0
build>=0.10.0
twine>=4.0.2

# Jupyter for development and analysis
jupyter>=1.0.0
ipython>=8.14.0
notebook>=7.0.0
